requirements.txt
openai>=1.0.0
chromadb>=0.4.0
pydantic>=2.0
python-dotenv>=1.2.1
tqdm